requirements.txt
numpy>=1.23
scipy>=1.9
pandas>=1.5
matplotlib>=3.6
pyyaml>=6.0
